requirements.txt
google-adk
streamlit
dotenv
fastapi
requests
pydantic
fastmcp
uvicorn
beautifulsoup4
langchain-tavily
langchain-openai
litellm